requirements.txt
opencv-python
numpy
pyyaml
pathlib
tk
pillow